[Style] feature/#172 랜딩페이지 완성 (#173)
* chore: 랜딩페이지 수정

* chore: 랜딩페이지 수정

* chore: 랜딩페이지 수정
