+++ src/pages/landing/LadingPage.jsx
@@ -2,6 +2,7 @@ import * as S from './LandingPage.style';
 
 import LANDING from '@/assets/images/landing.webp';
 import LANDING2 from '@/assets/images/landing2.webp';
+import { ServiceComponent } from '@/components';
 import Animation from '@/components/animation/Animation';
 
 const LadingPage = () => {
@@ -37,20 +38,25 @@ const LadingPage = () => {
 							우리 가족과의 소중한 시간을 <br />
 							놓치지 않게 관리하기
 						</h1>
+						<S.BrBox />
+						<h3>가족 채팅</h3>
+						<h1>
+							화목한 가족으로의 첫 걸음,
+							<br /> 대화로 먼저 시작하세요.
+						</h1>
 						<S.ImgBox>
 							<img src={LANDING2} alt="" />
 						</S.ImgBox>
 					</S.ContentWrapper>
 				</Animation>
 
-				<S.ContentWrapper>
-					<h3>가족 채팅</h3>
-					<h1>
-						화목한 가족으로의 첫 걸음,
-						<br /> 대화로 먼저 시작하세요.
-					</h1>
-				</S.ContentWrapper>
+				<Animation>
+					<S.ContentWrapper2>
+						<ServiceComponent />
+					</S.ContentWrapper2>
+				</Animation>
 			</S.ContentContainer>
+
 			<S.BackGround />
 			<S.BackGround2 />
 			<S.BackGround3 />
